% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_boxplot.R
\name{quick_boxplot}
\alias{quick_boxplot}
\title{boxplot quick function
this function creates a quick boxplot for visualizing
categorical and continuous variables}
\usage{
quick_boxplot(data, x, y, na.rm = FALSE)
}
\value{
a tidy, simple boxplot
}
\description{
boxplot quick function
this function creates a quick boxplot for visualizing
categorical and continuous variables
}
\details{
the function will keep NA values

the function will be tested for the creation and printing of the object, as well as correct axis labels
}
\examples{
quick_boxplot(cancer_sample, diagnosis, perimeter_mean)
quick_boxplot(cancer_sample, diagnosis, symmetry_mean)
quick_boxplot(cancer_sample, diagnosis, area_mean)
}
