% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_boxplot.R
\name{quick_boxplot}
\alias{quick_boxplot}
\title{boxplot quick function}
\usage{
quick_boxplot(data, x, y, na.rm = FALSE)
}
\arguments{
\item{data}{is the cancer_sample dataset from datateachr}

\item{x}{is a categorical variable from column in dataset (diagonosis)}

\item{y}{is a continuous variable from column in dataset (select any continuous column)}

\item{na.rm}{will keep 'NA' values in the boxplot}
}
\value{
a tidy, simple boxplot
}
\description{
this function creates a quick boxplot for visualizing
categorical and continuous variables in ggplot2
}
\details{
the function will keep NA values

the function will be tested for the creation and printing of the object, as well as correct axis labels
}
\examples{
quick_boxplot(datateachr::cancer_sample, diagnosis, perimeter_mean)
quick_boxplot(datateachr::cancer_sample, diagnosis, symmetry_mean)
quick_boxplot(datateachr::cancer_sample, diagnosis, area_mean)
}
